requirements.txt
asgiref
async-timeout
certifi
charset-normalizer
coreapi
coreschema
Django
django-cors-headers
django-modeltranslation
django-phonenumber-field
django-redis
django-filter
django-resized
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
drf-yasg
idna
inflection
itypes
Jinja2
MarkupSafe
mysqlclient
packaging
phonenumbers
Pillow
PyJWT
python-decouple
pytz
redis
requests
ruamel.yaml
ruamel.yaml.clib
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
